Improve error message on test failures
By using `expect` and `assert_eq!` instead of `assert_matches!`, the
values of the expected and actual results will be printed in the panic
message making it a lot easier to debug.

Applying `#[track_caller]` to all the assertion functions makes the
panic messages point to the place where they were called, not to within
the assertion function itself. This is a lot more useful.